+++ src/main/java/com/example/trendlog/controller/AuthController.java
@@ -1,12 +1,13 @@
 package com.example.trendlog.controller;
 
-import com.example.trendlog.dto.auth.*;
+import com.example.trendlog.dto.request.auth.TokenRefreshRequest;
+import com.example.trendlog.dto.request.auth.UserLoginRequest;
+import com.example.trendlog.dto.request.auth.UserSignupRequest;
+import com.example.trendlog.dto.response.auth.LoginResponse;
+import com.example.trendlog.dto.response.auth.TokenRefreshResponse;
 import com.example.trendlog.global.docs.AuthApiDoc;
 import com.example.trendlog.global.dto.DataResponse;
 import com.example.trendlog.service.AuthService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
