--- src/main/java/com/example/trendlog/global/exception/.gitkeep
@@ -1,1 +1,0 @@
-.gitkeep
